Resolving circ dep by extracting logger
